Add module template.
