+++ samples/template/README.md
@@ -1,0 +1,3 @@
+# Savant module template
+
+...
