+++ samples/template/main.py
@@ -1,0 +1,6 @@
+from savant.entrypoint.main import main
+
+
+if __name__ == '__main__':
+    main('module.yml')
+    print('done')
